finish js_backend refactor into multiple files
