--- src/lib/crypto/kimchi_bindings/js/js_backend/bindings.js
@@ -1,328 +1,0 @@
-/* global plonk_wasm, caml_js_to_bool, caml_jsstring_of_string, caml_string_of_jsstring,
-   caml_bytes_of_uint8array, caml_bytes_to_uint8array
-*/
-
-// Provides: caml_bigint_256_of_decimal_string
-// Requires: plonk_wasm, caml_jsstring_of_string
-var caml_bigint_256_of_decimal_string = function (s) {
-    return plonk_wasm.caml_bigint_256_of_decimal_string(caml_jsstring_of_string(s));
-};
-
-// Provides: caml_bigint_256_num_limbs
-// Requires: plonk_wasm
-var caml_bigint_256_num_limbs = plonk_wasm.caml_bigint_256_num_limbs
-
-// Provides: caml_bigint_256_bytes_per_limb
-// Requires: plonk_wasm
-var caml_bigint_256_bytes_per_limb = plonk_wasm.caml_bigint_256_bytes_per_limb
-
-// Provides: caml_bigint_256_div
-// Requires: plonk_wasm
-var caml_bigint_256_div = plonk_wasm.caml_bigint_256_div
-
-// Provides: caml_bigint_256_compare
-// Requires: plonk_wasm
-var caml_bigint_256_compare = plonk_wasm.caml_bigint_256_compare
-
-// Provides: caml_bigint_256_to_string
-// Requires: plonk_wasm, caml_string_of_jsstring
-var caml_bigint_256_to_string = function(x) {
-    return caml_string_of_jsstring(plonk_wasm.caml_bigint_256_to_string(x));
-};
-
-// Provides: caml_bigint_256_test_bit
-// Requires: plonk_wasm, caml_js_to_bool
-var caml_bigint_256_test_bit = function(x, i) {
-    return caml_js_to_bool(plonk_wasm.caml_bigint_256_test_bit(x, i));
-};
-
-// Provides: caml_bigint_256_to_bytes
-// Requires: plonk_wasm, caml_bytes_of_uint8array
-var caml_bigint_256_to_bytes = function(x) {
-    return caml_bytes_of_uint8array(plonk_wasm.caml_bigint_256_to_bytes(x));
-};
-
-// Provides: caml_bigint_256_of_bytes
-// Requires: plonk_wasm, caml_bytes_to_uint8array
-var caml_bigint_256_of_bytes = function(ocaml_bytes) {
-    return plonk_wasm.caml_bigint_256_of_bytes(caml_bytes_to_uint8array(ocaml_bytes));
-};
-
-
-
-
-
-
-// Provides: caml_pasta_fp_copy
-var caml_pasta_fp_copy = function(x, y) {
-    x[0] = y[0];
-};
-
-// Provides: caml_pasta_fp_option
-var caml_pasta_fp_option = function(x) {
-    // We encode 'none' in JS as undefined
-    if (x === undefined) return 0; // None
-    return [0, x]; // Some(x)
-};
-
-// Provides: caml_pasta_fp_size_in_bits
-// Requires: plonk_wasm
-var caml_pasta_fp_size_in_bits = plonk_wasm.caml_pasta_fp_size_in_bits
-
-// Provides: caml_pasta_fp_size
-// Requires: plonk_wasm
-var caml_pasta_fp_size = plonk_wasm.caml_pasta_fp_size
-
-// Provides: caml_pasta_fp_add
-// Requires: plonk_wasm
-var caml_pasta_fp_add = plonk_wasm.caml_pasta_fp_add
-
-// Provides: caml_pasta_fp_sub
-// Requires: plonk_wasm
-var caml_pasta_fp_sub = plonk_wasm.caml_pasta_fp_sub
-
-// Provides: caml_pasta_fp_negate
-// Requires: plonk_wasm
-var caml_pasta_fp_negate = plonk_wasm.caml_pasta_fp_negate
-
-// Provides: caml_pasta_fp_mul
-// Requires: plonk_wasm
-var caml_pasta_fp_mul = plonk_wasm.caml_pasta_fp_mul
-
-// Provides: caml_pasta_fp_div
-// Requires: plonk_wasm
-var caml_pasta_fp_div = plonk_wasm.caml_pasta_fp_div
-
-// Provides: caml_pasta_fp_inv
-// Requires: plonk_wasm, caml_pasta_fp_option
-var caml_pasta_fp_inv = function(x) {
-    return caml_pasta_fp_option(plonk_wasm.caml_pasta_fp_inv(x));
-};
-
-// Provides: caml_pasta_fp_square
-// Requires: plonk_wasm
-var caml_pasta_fp_square = plonk_wasm.caml_pasta_fp_square
-
-// Provides: caml_pasta_fp_is_square
-// Requires: plonk_wasm, caml_js_to_bool
-var caml_pasta_fp_is_square = function(x) {
-    return caml_js_to_bool(plonk_wasm.caml_pasta_fp_is_square(x));
-};
-
-// Provides: caml_pasta_fp_sqrt
-// Requires: plonk_wasm, caml_pasta_fp_option
-var caml_pasta_fp_sqrt = function(x) {
-    return caml_pasta_fp_option(plonk_wasm.caml_pasta_fp_sqrt(x));
-};
-
-// Provides: caml_pasta_fp_of_int
-// Requires: plonk_wasm
-var caml_pasta_fp_of_int = plonk_wasm.caml_pasta_fp_of_int
-
-// Provides: caml_pasta_fp_to_string
-// Requires: plonk_wasm, caml_string_of_jsstring
-var caml_pasta_fp_to_string = function(x) {
-    return caml_string_of_jsstring(plonk_wasm.caml_pasta_fp_to_string(x));
-};
-
-// Provides: caml_pasta_fp_of_string
-// Requires: plonk_wasm, caml_jsstring_of_string
-var caml_pasta_fp_of_string = function (x) {
-    return plonk_wasm.caml_pasta_fp_of_string(caml_jsstring_of_string(x));
-};
-
-
-// Provides: caml_pasta_fp_mut_add
-// Requires: caml_pasta_fp_copy, caml_pasta_fp_add
-var caml_pasta_fp_mut_add = function(x, y) {
-    caml_pasta_fp_copy(x, caml_pasta_fp_add(x, y));
-};
-
-// Provides: caml_pasta_fp_mut_sub
-// Requires: caml_pasta_fp_copy, caml_pasta_fp_sub
-var caml_pasta_fp_mut_sub = function(x, y) {
-    caml_pasta_fp_copy(x, caml_pasta_fp_sub(x, y));
-};
-
-// Provides: caml_pasta_fp_mut_mul
-// Requires: caml_pasta_fp_copy, caml_pasta_fp_mul
-var caml_pasta_fp_mut_mul = function(x, y) {
-    caml_pasta_fp_copy(x, caml_pasta_fp_mul(x, y));
-};
-
-// Provides: caml_pasta_fp_mut_square
-// Requires: caml_pasta_fp_copy, caml_pasta_fp_square
-var caml_pasta_fp_mut_square = function(x) {
-    caml_pasta_fp_copy(x, caml_pasta_fp_square(x));
-};
-
-// Provides: caml_pasta_fp_compare
-// Requires: plonk_wasm
-var caml_pasta_fp_compare = plonk_wasm.caml_pasta_fp_compare
-
-// Provides: caml_pasta_fp_equal
-// Requires: plonk_wasm
-var caml_pasta_fp_equal = plonk_wasm.caml_pasta_fp_equal
-
-// Provides: caml_pasta_fp_to_bigint
-// Requires: plonk_wasm
-var caml_pasta_fp_to_bigint = plonk_wasm.caml_pasta_fp_to_bigint
-
-// Provides: caml_pasta_fp_of_bigint
-// Requires: plonk_wasm
-var caml_pasta_fp_of_bigint = plonk_wasm.caml_pasta_fp_of_bigint
-
-// Provides: caml_pasta_fp_random
-// Requires: plonk_wasm
-var caml_pasta_fp_random = plonk_wasm.caml_pasta_fp_random;
-
-// Provides: caml_pasta_fp_to_bytes
-// Requires: plonk_wasm, caml_bytes_of_uint8array
-var caml_pasta_fp_to_bytes = function(x) {
-    var res = plonk_wasm.caml_pasta_fp_to_bytes(x);
-    return caml_bytes_of_uint8array(plonk_wasm.caml_pasta_fp_to_bytes(x));
-};
-
-// Provides: caml_pasta_fp_of_bytes
-// Requires: plonk_wasm, caml_bytes_to_uint8array
-var caml_pasta_fp_of_bytes = function(ocaml_bytes) {
-    return plonk_wasm.caml_pasta_fp_of_bytes(caml_bytes_to_uint8array(ocaml_bytes));
-};
-
-
-
-
-// Provides: caml_pasta_fq_copy
-var caml_pasta_fq_copy = function(x, y) {
-    x[0] = y[0];
-};
-
-// Provides: caml_pasta_fq_option
-var caml_pasta_fq_option = function(x) {
-    // We encode 'none' in JS as undefined
-    if (x === undefined) return 0; // None
-    return [0, x]; // Some(x)
-};
-
-// Provides: caml_pasta_fq_size_in_bits
-// Requires: plonk_wasm
-var caml_pasta_fq_size_in_bits = plonk_wasm.caml_pasta_fq_size_in_bits
-
-// Provides: caml_pasta_fq_size
-// Requires: plonk_wasm
-var caml_pasta_fq_size = plonk_wasm.caml_pasta_fq_size
-
-// Provides: caml_pasta_fq_add
-// Requires: plonk_wasm
-var caml_pasta_fq_add = plonk_wasm.caml_pasta_fq_add
-
-// Provides: caml_pasta_fq_sub
-// Requires: plonk_wasm
-var caml_pasta_fq_sub = plonk_wasm.caml_pasta_fq_sub
-
-// Provides: caml_pasta_fq_negate
-// Requires: plonk_wasm
-var caml_pasta_fq_negate = plonk_wasm.caml_pasta_fq_negate
-
-// Provides: caml_pasta_fq_mul
-// Requires: plonk_wasm
-var caml_pasta_fq_mul = plonk_wasm.caml_pasta_fq_mul
-
-// Provides: caml_pasta_fq_div
-// Requires: plonk_wasm
-var caml_pasta_fq_div = plonk_wasm.caml_pasta_fq_div
-
-// Provides: caml_pasta_fq_inv
-// Requires: plonk_wasm, caml_pasta_fq_option
-var caml_pasta_fq_inv = function(x) {
-    return caml_pasta_fq_option(plonk_wasm.caml_pasta_fq_inv(x));
-};
-
-// Provides: caml_pasta_fq_square
-// Requires: plonk_wasm
-var caml_pasta_fq_square = plonk_wasm.caml_pasta_fq_square
-
-// Provides: caml_pasta_fq_is_square
-// Requires: plonk_wasm, caml_js_to_bool
-var caml_pasta_fq_is_square = function(x) {
-    return caml_js_to_bool(plonk_wasm.caml_pasta_fq_is_square(x));
-};
-
-// Provides: caml_pasta_fq_sqrt
-// Requires: plonk_wasm, caml_pasta_fq_option
-var caml_pasta_fq_sqrt = function(x) {
-    return caml_pasta_fq_option(plonk_wasm.caml_pasta_fq_sqrt(x));
-};
-
-// Provides: caml_pasta_fq_of_int
-// Requires: plonk_wasm
-var caml_pasta_fq_of_int = plonk_wasm.caml_pasta_fq_of_int
-
-// Provides: caml_pasta_fq_to_string
-// Requires: plonk_wasm, caml_string_of_jsstring
-var caml_pasta_fq_to_string = function(x) {
-    return caml_string_of_jsstring(plonk_wasm.caml_pasta_fq_to_string(x));
-};
-
-// Provides: caml_pasta_fq_of_string
-// Requires: plonk_wasm, caml_jsstring_of_string
-var caml_pasta_fq_of_string = function (x) {
-    return plonk_wasm.caml_pasta_fq_of_string(caml_jsstring_of_string(x));
-};
-
-// Provides: caml_pasta_fq_mut_add
-// Requires: caml_pasta_fq_copy, caml_pasta_fq_add
-var caml_pasta_fq_mut_add = function(x, y) {
-    caml_pasta_fq_copy(x, caml_pasta_fq_add(x, y));
-};
-
-// Provides: caml_pasta_fq_mut_sub
-// Requires: caml_pasta_fq_copy, caml_pasta_fq_sub
-var caml_pasta_fq_mut_sub = function(x, y) {
-    caml_pasta_fq_copy(x, caml_pasta_fq_sub(x, y));
-};
-
-// Provides: caml_pasta_fq_mut_mul
-// Requires: caml_pasta_fq_copy, caml_pasta_fq_mul
-var caml_pasta_fq_mut_mul = function(x, y) {
-    caml_pasta_fq_copy(x, caml_pasta_fq_mul(x, y));
-};
-
-// Provides: caml_pasta_fq_mut_square
-// Requires: caml_pasta_fq_copy, caml_pasta_fq_square
-var caml_pasta_fq_mut_square = function(x) {
-    caml_pasta_fq_copy(x, caml_pasta_fq_square(x));
-};
-
-// Provides: caml_pasta_fq_compare
-// Requires: plonk_wasm
-var caml_pasta_fq_compare = plonk_wasm.caml_pasta_fq_compare
-
-// Provides: caml_pasta_fq_equal
-// Requires: plonk_wasm
-var caml_pasta_fq_equal = plonk_wasm.caml_pasta_fq_equal
-
-// Provides: caml_pasta_fq_to_bigint
-// Requires: plonk_wasm
-var caml_pasta_fq_to_bigint = plonk_wasm.caml_pasta_fq_to_bigint
-
-// Provides: caml_pasta_fq_of_bigint
-// Requires: plonk_wasm
-var caml_pasta_fq_of_bigint = plonk_wasm.caml_pasta_fq_of_bigint
-
-// Provides: caml_pasta_fq_random
-// Requires: plonk_wasm
-var caml_pasta_fq_random = plonk_wasm.caml_pasta_fq_random;
-
-// Provides: caml_pasta_fq_to_bytes
-// Requires: plonk_wasm, caml_bytes_of_uint8array
-var caml_pasta_fq_to_bytes = function(x) {
-    var res = plonk_wasm.caml_pasta_fq_to_bytes(x);
-    return caml_bytes_of_uint8array(plonk_wasm.caml_pasta_fq_to_bytes(x));
-};
-
-// Provides: caml_pasta_fq_of_bytes
-// Requires: plonk_wasm, caml_bytes_to_uint8array
-var caml_pasta_fq_of_bytes = function(ocaml_bytes) {
-    return plonk_wasm.caml_pasta_fq_of_bytes(caml_bytes_to_uint8array(ocaml_bytes));
-};
