+++ src/lib/crypto/kimchi_bindings/js/js_backend/dune
@@ -2,9 +2,9 @@
  (name js_backend)
  (js_of_ocaml
   (javascript_files
-    js_backend.js
-    bindings.js
+    finite_field.js
     elliptic_curve.js
+    bigint256.js
     bigint_helpers.js
     vector.js
     unimplemented.js
